#6153, Improve the routing when having multi table index in one model. (#6236)
#6153, Improve the routing when having multi table index in one model.

+++ core-metadata/src/main/java/org/apache/kylin/metadata/realization/CapabilityResult.java
@@ -37,7 +37,7 @@ public class CapabilityResult {
     /**
      * The smaller the cost, the more capable the realization
      */
-    public int cost;
+    public double cost;
 
     /**
      * reason of incapable
