+++ core-storage/src/main/java/org/apache/kylin/storage/hybrid/HybridInstance.java
@@ -219,7 +219,7 @@ public class HybridInstance extends RootPersistentEntity implements IRealization
     }
 
     @Override
-    public int getCost(SQLDigest sqlDigest) {
+    public double getCost(SQLDigest sqlDigest) {
         return getCost();
     }
